requirements.txt
Django>=3.1.6,<3.2.0
djangorestframework>=3.12.2,<3.13.0

redis>=3.2.100,<3.3.0

python-dotenv>=0.15.0,<0.16.0
